+++ packages/blade/src/components/Button/BaseButton/__tests__/BaseButton.ssr.test.tsx
@@ -1,0 +1,15 @@
+import BaseButton from '../BaseButton';
+import renderWithSSR from '~src/_helpers/testing/renderWithSSR.web';
+import { CreditCardIcon } from '~components/Icons';
+
+describe('<BaseButton />', () => {
+  it('should render button with icon with left iconPosition', () => {
+    const buttonText = 'Pay Now';
+    const { container } = renderWithSSR(
+      <BaseButton iconPosition="left" icon={CreditCardIcon}>
+        {buttonText}
+      </BaseButton>,
+    );
+    expect(container).toMatchSnapshot();
+  });
+});