+++ packages/blade/src/components/index.ts
@@ -1,6 +1,7 @@
 export * from './Alert';
 export * from './Badge';
 export * from './BladeProvider';
+export * from './Card';
 export * from './Button';
 export * from './Button/IconButton';
 export * from './Counter';
@@ -13,6 +14,7 @@ export * from './Input/OTPInput';
 export * from './Indicator';
 export * from './Link';
 export * from './LiveAnnouncer';
+export * from './ProgressBar';
 export * from './Radio';
 export * from './Spinner';
 export * from './SkipNav';