+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @razorpay/blade
 
+## 5.4.0
+
+### Minor Changes
+
+- 5c8005f: feat: add `ProgressBar` component
+
+## 5.3.0
+
+### Minor Changes
+
+- 26baa81: feat(blade): added Card component
+
 ## 5.2.1
 
 ### Patch Changes